backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.2
google-api-python-client==2.149.0
google-auth==2.35.0
oauth2client<4.0.0
pydantic==2.9.2
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
streamlit==1.39.0
httpx==0.27.2
pytest==8.3.3
prometheus-client==0.21.0
redis==5.0.8
openai==1.51.2